Лучшее, part1
Переключение в основную активность с журнала

indicator

ThumborUrlBuilder.stripicc(), build

Logimage download

subscribtionsActivity

продолжаем

SubscribtionsActivity

Полупрозрачныые бары на api19 при просмотре фото

SubscriptionHelper.empty();

MyFeedActivity, новости

indicator

Продолжаем

+++ app/src/main/java/ru/taaasty/service/ApiTlog.java
@@ -18,7 +18,7 @@ public interface ApiTlog {
 
     @GET("/tlog/{id}/entries.json")
     Observable<Feed> getEntries(@Path("id") String idOrSlug,
-                                @Query("since_entry_id") Integer sinceEntryId,
+                                @Query("since_entry_id") Long sinceEntryId,
                                 @Query("limit") Integer limit);
 
     // XXX
